fix concurrency
